[fix]게시글 첨부파일 조회

+++ src/main/java/com/spring/boot/common/FileUtils.java
@@ -15,7 +15,7 @@ public class FileUtils {
 	@Value("${fileUpload.path}")
 	private String path;
 	
-	public FileInfoDto.request parseFileInfo(MultipartFile file, int memberId){
+	public FileInfoDto.request parseFileInfo(MultipartFile file, long memberId){
 		File target = new File(path);
 		if(!target.exists()) {
 			target.mkdirs();
